Zsh: add highlighter package
I'm not sure I'm going to keep this.

It has a tone of features, grouped by course types.  I like the
`bracket' group, which is what drove me to set this up.  I like *some*
of the things in `main', but not all of it.  I'm not ready to invest the
time required to figure out how to deactivate those unwanted parts.

https://github.com/zsh-users/zsh-syntax-highlighting/blob/master/docs/highlighters/main.md